requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.11.0
jwt==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
